data_pipeline/src/requirements.txt
APScheduler
confluent_kafka
flask
psycopg
psycopg-binary
pydantic
vertica_python
apache-airflow-providers-vertica
apache-airflow-providers-postgres